+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "node src/server/database/models/__tests__/Movie.spec.js",
+    "test": "node src/server/models/movie/__tests__/Movie.spec.js",
     "start": "node dist/server/index.js",
     "build": "npm run build:server",
     "build:server": "babel src/server -d dist/server"
